Restructure module dependencies of v1/v2 seq_mse
